refactor repo structure, rm browser dir
