+++ dcp_prototype/backend/code/common/browser_orm.py
@@ -15,10 +15,10 @@ from sqlalchemy import (
 from sqlalchemy.orm import relationship, sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
 
-pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../.."))  # noqa
+pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from browser.code.config.db_config import BrowserDbConfig
+from code.config.db_config import BrowserDbConfig
 
 
 Base = declarative_base()
