+++ dcp_prototype/backend/code/common/db_utils.py
@@ -2,10 +2,10 @@ import os
 import sys
 import typing
 
-pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../../.."))  # noqa
+pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from browser.code.common.browser_orm import (
+from code.common.browser_orm import (
     Base,
     DBSessionMaker,
     Project,
