+++ dcp_prototype/backend/iam/policy-templates/browser-api-lambda.json
