+++ dcp_prototype/backend/scripts/create_db.py
@@ -3,11 +3,11 @@ import sys
 
 from sqlalchemy import create_engine
 
-pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), "../.."))  # noqa
+pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from browser.code.config.db_config import BrowserDbConfig
-from browser.code.common import browser_orm as orm
+from code.config.db_config import BrowserDbConfig
+from code.common import browser_orm as orm
 
 
 if __name__ == "__main__":
