+++ dcp_prototype/backend/scripts/generate_test_artifact.py
