+++ dcp_prototype/frontend/README.md
@@ -25,7 +25,7 @@
     
 1.  **Host the backend locally.**
 
-    Follow [backend instructions](https://github.com/chanzuckerberg/dcp-prototype/tree/master/dcp_prototype/backend/browser#development)
+    Follow [backend instructions](https://github.com/chanzuckerberg/dcp-prototype/tree/master/dcp_prototype/backend#development)
     to deploy the backend API on `http://localhost:5000`.
 
 1.  **Build and launch the frontend locally.**
