--- tests/dcp-prototype/__init__.py
@@ -1,4 +1,0 @@
-import os
-
-os.environ["DEPLOYMENT_STAGE"] = "test"
-os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
