+++ tests/unit/__init__.py
@@ -1,3 +1,4 @@
 import os
 
 os.environ["DEPLOYMENT_STAGE"] = "test"
+os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
