+++ tests/unit/dcp-prototype/backend/code/common/test_db_utils.py
@@ -2,15 +2,15 @@ import json
 import unittest
 from unittest import mock
 
-from dcp_prototype.backend.browser.code.common.db_utils import DbUtils
-from dcp_prototype.backend.browser.scripts.load_artifact import load_from_artifact
-from dcp_prototype.backend.browser.scripts.mock import mock_data
+from dcp_prototype.backend.code.common.db_utils import DbUtils
+from dcp_prototype.backend.scripts.load_artifact import load_from_artifact
+from dcp_prototype.backend.scripts.mock import mock_data
 
 
 class TestDbUtils(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        path_to_fixtures = "tests/unit/dcp-prototype/backend/browser/fixtures"
+        path_to_fixtures = "tests/unit/dcp-prototype/backend/fixtures"
         test_artifact_file = f"{path_to_fixtures}/test_artifact.json"
         with open(test_artifact_file, "r") as f:
             cls.test_artifact = json.loads(f.read())
@@ -40,7 +40,7 @@ class TestDbUtils(unittest.TestCase):
 
         with self.subTest("Throws EnvironmentError if not test db"):
             with mock.patch(
-                "dcp_prototype.backend.browser.code.common.db_utils.DbUtils._is_test_db"
+                "dcp_prototype.backend.code.common.db_utils.DbUtils._is_test_db"
             ) as mock_is_test_db:
                 mock_is_test_db.return_value = False
                 with self.assertRaises(EnvironmentError):
